Update myBoards.cpp and Docs
There was issue with motor jerk on setup() during init of the driver board code. Moving the init of the inputPins[] to before the driver board init appears to be a solution.
Have also updated docs for hoefully the last time. That is everything from me I think fore release 1. I have changed Version number of firmware to 100,